requirements.txt
python-dotenv 
langchain 
langchain-openai 
streamlit 
langchain-community